+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.0.0 - Combinator selector replacement
+* combinator selectors that can't be prefixed with the prefix are replaced with the prefix.
+* dependencies update.
+
 ## 1.0.1 - Fix for node 0.10 & 0.12
 * fixed TypeError: Object has no method 'includes'.